requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
scikit-learn>=1.3
statsmodels>=0.14
pmdarima>=2.0
streamlit>=1.36
plotly>=5.22
pyyaml>=6.0
